+++ src/core/server/systems/tick.ts
@@ -46,4 +46,11 @@ function handlePing(player: alt.Player): void {
     if (player.nextDeathSpawn && Date.now() > player.nextDeathSpawn - 1000) {
         playerFuncs.set.respawned(player, null); // Uses null to find a hospital.
     }
+
+    // Updates Food & Water
+    if (!player.nextFoodSync || Date.now() > player.nextFoodSync) {
+        player.nextFoodSync = Date.now() + DEFAULT_CONFIG.TIME_BETWEEN_FOOD_UPDATES;
+        playerFuncs.sync.food(player);
+        playerFuncs.sync.water(player);
+    }
 }